requirements.txt
autopep8==1.7.0
click==8.1.3
dnspython==2.2.1
Flask==2.2.2
Flask-PyMongo==2.3.0
itsdangerous==2.1.2
Jinja2==3.1.2
livereload==2.6.3
MarkupSafe==2.1.1
pycodestyle==2.9.1
pymongo==4.2.0
python-dotenv==0.21.0
six==1.16.0
toml==0.10.2
tornado==6.2
Werkzeug==2.2.2
